Move `dup` script to python
There was a problem when trying to use duplicity's `--include`, `--exclude` and similar tools, and using wildcards there.

The current eval + echo system was evaluating twice the output. Example:

    OPTIONS="--include /mnt/backup/src/volumes/*_*"
    backup
    [...]
    Command line error: Expected 2 args, got 7

It's quite tricky to fix with bash, and even trickier with busybox's shell, so at the end, I decided to use a real programming language instead: Python.

Now, `$OPTIONS` and `$OPTIONS_EXTRA` are guaranteed to be evaluated only once.