Update requests dependency (removing mediafire backend)
The Requests package through 2.19.1 before 2018-09-14 for Python sends an HTTP Authorization header to an http URI upon receiving a same-hostname https-to-http redirect, which makes it easier for remote attackers to discover credentials by sniffing the network. (CVE-2018-18074)

We should abandon that old package in favour of a most recent one.

We need to remove the mediafire backend to do so.